requirements.txt
# audio
sounddevice
soundfile
webrtcvad

# stt
faster-whisper

# llm
requests

# tts
edge-tts

# vts websocket
websocket-client
